+++ elenasrc2/elc/derivation.cpp
@@ -735,7 +735,7 @@ void DerivationWriter :: recognizeClassMebers(SNode node/*, DerivationScope& sco
    SNode current = node.firstChild();
    while (current != lxNone) {
       if (current == lxScope) {
-         SNode bodyNode = current.findChild(lxCode, lxDispatchCode, lxReturning, lxExpression, lxResendExpression);
+         SNode bodyNode = current.findChild(lxCode, lxDispatchCode, lxReturning, lxExpression, lxResendExpression, lxNoBody);
 
          int mode = 0;
          if (bodyNode == lxExpression) {
@@ -1302,7 +1302,7 @@ void DerivationWriter :: generateMethodTree(SyntaxWriter& writer, SNode node, Sc
       writer.closeNode();
    }
    else {
-      SNode bodyNode = node.findChild(lxCode, lxDispatchCode, lxReturning, lxResendExpression);
+      SNode bodyNode = node.findChild(lxCode, lxDispatchCode, lxReturning, lxResendExpression, lxNoBody);
       if (bodyNode.compare(lxReturning, lxDispatchCode)) {
          writer.newNode(bodyNode.type);
          generateExpressionTree(writer, bodyNode.firstChild(), derivationScope, EXPRESSION_IMPLICIT_MODE);
@@ -1320,6 +1320,9 @@ void DerivationWriter :: generateMethodTree(SyntaxWriter& writer, SNode node, Sc
       }
       else if (bodyNode == lxCode) {
          generateCodeTree(writer, bodyNode, derivationScope);
+      }
+      else if (bodyNode == lxNoBody) {
+         writer.appendNode(lxNoBody);
       }
    }